+++ proxy/https.go
@@ -45,8 +45,8 @@ func (pxy *Proxy) handleHttps(lConn *net.TCPConn, exploit bool, initPkt *packet.
 	log.Debug("[HTTPS] Sent 200 Connection Estabalished to ", lConn.RemoteAddr())
 
 	// Read client hello
-	m, err := ReadTlsMessage(lConn)
-	if err != nil || !IsClientHello(m) {
+	m, err := packet.ReadTLSMessage(lConn)
+	if err != nil || !m.IsClientHello() {
 		log.Debug("[HTTPS] Error reading client hello from the client", err)
 		return
 	}
@@ -55,24 +55,18 @@ func (pxy *Proxy) handleHttps(lConn *net.TCPConn, exploit bool, initPkt *packet.
 	log.Debug("[HTTPS] Client sent hello ", len(clientHello), "bytes")
 
 	// Generate a go routine that reads from the server
-
-	chPkt := packet.NewHttpsPacket(clientHello)
-
-	// lConn.SetLinger(3)
-	// rConn.SetLinger(3)
-
 	go Serve(rConn, lConn, "[HTTPS]", rConn.RemoteAddr().String(), initPkt.Domain(), pxy.timeout, pxy.bufferSize)
 
 	if exploit {
 		log.Debug("[HTTPS] Writing chunked client hello to ", initPkt.Domain())
-		chunks := splitInChunks(chPkt.Raw(), pxy.windowSize)
-		if _, err := WriteChunks(rConn, chunks); err != nil {
+		chunks := splitInChunks(clientHello, pxy.windowSize)
+		if _, err := writeChunks(rConn, chunks); err != nil {
 			log.Debug("[HTTPS] Error writing chunked client hello to ", initPkt.Domain(), err)
 			return
 		}
 	} else {
 		log.Debug("[HTTPS] Writing plain client hello to ", initPkt.Domain())
-		if _, err := rConn.Write(chPkt.Raw()); err != nil {
+		if _, err := rConn.Write(clientHello); err != nil {
 			log.Debug("[HTTPS] Error writing plain client hello to ", initPkt.Domain(), err)
 			return
 		}
@@ -116,3 +110,18 @@ func splitInChunks(bytes []byte, size int) [][]byte {
 
 	return [][]byte{raw[:1], raw[1:]}
 }
+
+func writeChunks(conn *net.TCPConn, c [][]byte) (n int, err error) {
+	total := 0
+	for i := 0; i < len(c); i++ {
+		b, err := conn.Write(c[i])
+		if err != nil {
+			return 0, nil
+		}
+
+		total += b
+	}
+
+	return total, nil
+}
+
